fix problem and optimization
